+++ lib/process_bot/process.rb
@@ -12,6 +12,7 @@ class ProcessBot::Process
 
   autoload :Handlers, "#{__dir__}/process/handlers"
   autoload :Runner, "#{__dir__}/process/runner"
+  autoload :RunnerInstance, "#{__dir__}/process/runner_instance"
 
   attr_reader :control_command_monitor, :current_pid, :current_process_title, :options, :port, :stopped
 
@@ -21,6 +22,9 @@ class ProcessBot::Process
     @accept_control_commands = true
     @control_command_monitor = Monitor.new
     @control_commands_in_flight = 0
+    @runner_events = Queue.new
+    @runner_instances = []
+    @runner_monitor = Monitor.new
 
     options.events.connect(:on_process_started, &method(:on_process_started)) # rubocop:disable Performance/MethodObjectAsBlock
     options.events.connect(:on_socket_opened, &method(:on_socket_opened)) # rubocop:disable Performance/MethodObjectAsBlock
@@ -34,7 +38,7 @@ class ProcessBot::Process
     if command == "start"
       logger.logs "Starting process"
       start
-    elsif command == "graceful" || command == "graceful_no_wait" || command == "stop"
+    elsif command == "graceful" || command == "graceful_no_wait" || command == "restart" || command == "stop"
       send_control_command(command)
     else
       raise "Unknown command: #{command}"
@@ -85,18 +89,32 @@ class ProcessBot::Process
 
   def start
     start_control_socket
+    start_runner_instance
 
     loop do
-      run
+      runner_event = runner_events.pop
+      handle_runner_event(runner_event)
 
-      if stopped
-        stop_accepting_control_commands
-        wait_for_control_commands
-        break
+      next unless stopped && runner_instances.empty?
+
+      stop_accepting_control_commands
+      wait_for_control_commands
+      break
+    end
+  end
+
+  def restart(**_args)
+    logger.logs "Restart process"
+
+    if handler_name == "sidekiq"
+      if restart_overlap?
+        handler_instance.graceful_no_wait(stop_process_bot: false)
+        start_runner_instance
       else
-        logger.logs "Process stopped - starting again after 1 sec"
-        sleep 1
+        handler_instance.graceful(stop_process_bot: false)
       end
+    else
+      handler_instance.stop
     end
   end
 
@@ -107,7 +125,7 @@ class ProcessBot::Process
   end
 
   def run
-    runner.run
+    start_runner_instance
   end
 
   def send_control_command(command, **command_options)
@@ -119,13 +137,7 @@ class ProcessBot::Process
   end
 
   def runner
-    @runner ||= ProcessBot::Process::Runner.new(
-      command: handler_instance.start_command,
-      handler_name: handler_name,
-      handler_instance: handler_instance,
-      logger: logger,
-      options: options
-    )
+    current_runner_instance&.runner || @runner ||= build_runner
   end
 
   def update_process_title
@@ -164,4 +176,90 @@ class ProcessBot::Process
       @control_commands_in_flight
     end
   end
+
+  def build_runner
+    ProcessBot::Process::Runner.new(
+      command: handler_instance.start_command,
+      handler_name: handler_name,
+      handler_instance: handler_instance,
+      logger: logger,
+      options: options
+    )
+  end
+
+  def start_runner_instance
+    runner_instance = ProcessBot::Process::RunnerInstance.new(
+      runner: build_runner,
+      event_queue: runner_events,
+      logger: logger
+    )
+
+    track_runner_instance(runner_instance)
+    @current_runner_instance = runner_instance
+    @runner = runner_instance.runner
+    runner_instance.start
+  end
+
+  def handle_runner_event(runner_event)
+    runner_instance = runner_event.fetch(:runner_instance)
+    remove_runner_instance(runner_instance)
+    log_runner_event_error(runner_event)
+    clear_current_runner(runner_instance)
+    restart_runner_if_needed(runner_instance)
+  end
+
+  def runner_instances
+    runner_monitor.synchronize do
+      @runner_instances.dup
+    end
+  end
+
+  def track_runner_instance(runner_instance)
+    runner_monitor.synchronize do
+      @runner_instances << runner_instance
+    end
+  end
+
+  def remove_runner_instance(runner_instance)
+    runner_monitor.synchronize do
+      @runner_instances.delete(runner_instance)
+    end
+  end
+
+  def restart_overlap?
+    value = options[:sidekiq_restart_overlap]
+    return false if value.nil?
+    return value if value == true || value == false
+
+    normalized = value.to_s.strip.downcase
+    return false if normalized == "false" || normalized == "0" || normalized == ""
+
+    true
+  end
+
+  def log_runner_event_error(runner_event)
+    return unless runner_event.fetch(:type) == :error
+
+    logger.error "Process runner crashed: #{runner_event.fetch(:error)}"
+  end
+
+  def clear_current_runner(runner_instance)
+    return unless runner_instance == current_runner_instance
+
+    @current_runner_instance = nil
+    @runner = nil
+  end
+
+  def restart_runner_if_needed(runner_instance)
+    return if stopped
+    return unless runner_instance == current_runner_instance || current_runner_instance.nil?
+
+    logger.logs "Process stopped - starting again after 1 sec"
+    sleep 1
+    start_runner_instance
+  end
+
+private
+
+  attr_reader :current_runner_instance, :runner_events, :runner_monitor
 end